soildetection/requirements.txt
torch
torchvision
streamlit
pillow
PyYAML
tqdm
matplotlib
